+++ lib/commands/init_command.dart
@@ -11,14 +11,14 @@ class InitCommand extends Command {
   @override
   void run() async {
     /// Creating assets_cleaner.yaml file
-    final result = await ConfigServices.instance.createConfigFile();
+    final result = await ConfigServices.instance.create();
     if (result) {
       print("---------------------------------------");
       print("[✅] Successfully create assets_cleaner.yaml file");
       print("---------------------------------------");
     } else {
       print("---------------------------------------");
-      print("[❌] Failed to create assets_cleaner.yaml file");
+      print("[🚫] Failed to create assets_cleaner.yaml file");
       print("---------------------------------------");
     }
   }