+++ lib/commands/unused_command.dart
@@ -1,23 +1,27 @@
-
 import 'package:args/command_runner.dart';
 import 'package:assets_cleaner/models/unused_asset_model.dart';
 import 'package:assets_cleaner/services/asset_services.dart';
 
 class UnusedCommand extends Command {
-
   @override
-  String get description => 'Scanning unused assets in lib folder and display it';
+  String get description =>
+      'Scanning unused assets in lib folder and display it';
 
   @override
   String get name => 'unused';
 
-
   @override
   void run() async {
-    /// Get list of unused assets from 
+    /// Get list of unused assets from
     /// lib folder and read .dart file code
-    List<UnusedAssetModel> unusedAssets = await AssetServices.instance.getUnusedAssets();
-    if (unusedAssets.isEmpty) return;
+    List<UnusedAssetModel> unusedAssets =
+        await AssetServices.instance.getUnusedAssets();
+    if (unusedAssets.isEmpty) {
+      print("-----------------------------");
+      print("[✅] No unused assets found");
+      print("-----------------------------");
+      return;
+    }
 
     print("-----------------------------");
     for (var unused in unusedAssets) {