+++ lib/services/asset_services.dart
@@ -1,10 +1,8 @@
-
 import 'package:assets_cleaner/models/unused_asset_model.dart';
 import 'package:assets_cleaner/utils/code_utils.dart';
 import 'package:assets_cleaner/utils/file_utils.dart';
 
 class AssetServices {
-
   factory AssetServices() => instance;
   AssetServices._();
   static final AssetServices instance = AssetServices._();
@@ -12,13 +10,13 @@ class AssetServices {
   /// Get list of unused assets from
   /// lib folder and read .dart file code
   Future<List<UnusedAssetModel>> getUnusedAssets() async {
-    String currentPath = FileUtils.instance.getCurrentPath();
+    String currentPath = FileUtils.instance.getCurrentPath;
     final fileUtils = FileUtils();
 
     /// Read assets path from pubspec.yaml
     List<String> assetsPath = await fileUtils.getPubspecAsset();
     if (assetsPath.isEmpty) return [];
-    
+
     /// Get config from assets_cleaner.yaml
     fileUtils.getConfig();
     if (fileUtils.config == null) return [];
@@ -36,13 +34,14 @@ class AssetServices {
     /// Get exclude file and extensions
     final excludeFile = fileUtils.getExcludeFile();
     final excludeExtension = fileUtils.getExcludeExtension();
+
     /// Remove exclude extension and file from assets
     assets = _removeExcludedExt(excludeExtension, assets);
     assets = _removeExcludedFile(excludeFile, assets);
-    
+
     List<UnusedAssetModel> unusedAssets = [];
-    if (assets.isEmpty) return[];
-    
+    if (assets.isEmpty) return [];
+
     print("-----------------------------");
     print("Scanning ${assets.length} assets file");
     for (var asset in assets) {
@@ -55,7 +54,10 @@ class AssetServices {
   }
 
   /// Remove exclude extension from assets
-  List<String> _removeExcludedExt(List<String> excludedExtensions, List<String> value) {
+  List<String> _removeExcludedExt(
+    List<String> excludedExtensions,
+    List<String> value,
+  ) {
     for (var ext in excludedExtensions) {
       value.removeWhere((item) => item.contains(".$ext"));
     }
@@ -63,7 +65,10 @@ class AssetServices {
   }
 
   /// Remove exclude file from assets
-  List<String> _removeExcludedFile(List<String> excludedFiles, List<String> value) {
+  List<String> _removeExcludedFile(
+    List<String> excludedFiles,
+    List<String> value,
+  ) {
     for (var file in excludedFiles) {
       value.removeWhere((item) => item == file);
     }