+++ lib/services/config_services.dart
@@ -5,7 +5,7 @@ class ConfigServices {
   ConfigServices._();
   static final ConfigServices instance = ConfigServices._();
 
-  Future<bool> createConfigFile() async {
+  Future<bool> create() async {
     try {
       String templateUrl = "https://raw.githubusercontent.com/yusriltakeuchi/assets_cleaner/master/template/assets_cleaner.yaml";
       /// Downloading from the url and save to [path]